Added Bean validation
Also converted price and quantity to big double as bean validation does not allow @Min and @Max on doubles

+++ src/main/java/com/example/javaproject/Trade.java
@@ -13,17 +13,17 @@ import java.util.Date;
 public class Trade {
     private Order oldOrder;
     private  Order newOrder;
-    private float price;
-    private int quantity;
+    private double price;
+    private double quantity;
     private Instant date = Instant.now();
 
     public Trade(Order oldOrder, Order newOrder) {
         this.oldOrder = oldOrder;
         this.newOrder = newOrder;
-        this.price = oldOrder.getPrice();
-        int quantity = newOrder.getQuantity();
-        if(oldOrder.getQuantity()< newOrder.getQuantity()){
-            quantity = oldOrder.getQuantity();
+        this.price = oldOrder.getPrice().doubleValue();
+        double quantity = newOrder.getQuantity().doubleValue();
+        if(oldOrder.getQuantity().doubleValue()< newOrder.getQuantity().doubleValue()){
+            quantity = oldOrder.getQuantity().doubleValue();
         }
         this.quantity = quantity;
     }
